number
number

+++ MOFSPickerManagerDemo/MOFSPickerManager/MOFSPickerView.m
@@ -107,7 +107,7 @@
     self.toolBar.commitBlock = ^ {
         [weakSelf hiddenWithAnimation];
         if (commitBlock) {
-            NSString *rowStr = [NSString stringWithFormat:@"%lu",weakSelf.selectedRow];
+            NSString *rowStr = [NSString stringWithFormat:@"%ld",(long)weakSelf.selectedRow];
             [weakSelf.recordDic setValue:rowStr forKey:tagStr];
             commitBlock(weakSelf.dataArr[weakSelf.selectedRow]);
         }
@@ -124,7 +124,7 @@
     
     [self reloadAllComponents];
     self.selectedRow = 0;
-    NSString *tagStr = [NSString stringWithFormat:@"%lu",self.showTag];
+    NSString *tagStr = [NSString stringWithFormat:@"%ld",(long)self.showTag];
     if ([self.recordDic.allKeys containsObject:tagStr]) {
         self.selectedRow = [self.recordDic[tagStr] integerValue];
     }
@@ -143,7 +143,7 @@
     self.toolBar.commitBlock = ^ {
         [weakSelf hiddenWithAnimation];
         if (commitBlock) {
-            NSString *rowStr = [NSString stringWithFormat:@"%zd",weakSelf.selectedRow];
+            NSString *rowStr = [NSString stringWithFormat:@"%ld",(long)weakSelf.selectedRow];
             [weakSelf.recordDic setValue:rowStr forKey:tagStr];
             commitBlock(weakSelf.dataArr[weakSelf.selectedRow]);
         }
